+++ components/parking-lot/ConfigPlan.tsx
@@ -10,7 +10,7 @@ import SubHeaderText from "../text/SubHeaderText";
 import BodyText from "@/components/text/BodyText";
 import Colors from "@/constants/color";
 import { InputType } from "@/enum/InputType";
-import { ParkingLotRequest, Plan } from "@/types/parking-lot/ParkingLot";
+import { ParkingLotRequest, Plan } from "@/types/parking-lot";
 import ParkingZoneInput from "../input/ParkingZoneInput";
 import { ImageProps } from "@/types";
